+++ src/server/namespace.go
@@ -7,8 +7,8 @@ import (
 
 type Namespace interface {
 	Create(ctx context.Context, name string) (model.Namespace, error)
-	Update(ctx context.Context, uid uint64, name string) (model.Namespace, error)
-	Delete(ctx context.Context, uid uint64) error
-	GetList(ctx context.Context, limit, offset uint64) ([]model.Namespace, error)
+	Update(ctx context.Context, uid uint, name string) (model.Namespace, error)
+	Delete(ctx context.Context, uid uint) error
+	GetList(ctx context.Context, limit, offset uint) ([]model.Namespace, error)
 	GetByName(ctx context.Context, namespace string) (model.Namespace, error)
 }
