+++ src/server/repository/tag.go
@@ -7,9 +7,9 @@ import (
 
 type Tag interface {
 	Create(ctx context.Context, data *model.TagData) (model.Tag, error)
-	Update(ctx context.Context, id uint64, data *model.TagData) (model.Tag, error)
-	DeleteById(ctx context.Context, id uint64) error
-	GetById(ctx context.Context, id uint64) (model.Tag, error)
+	Update(ctx context.Context, id uint, data *model.TagData) (model.Tag, error)
+	DeleteById(ctx context.Context, id uint) error
+	GetById(ctx context.Context, id uint) (model.Tag, error)
 	GetByName(ctx context.Context, name string) (model.Tag, error)
-	GetByFilter(ctx context.Context, filter model.TagFilter, limit, offset uint64) ([]model.Tag, error)
+	GetByFilter(ctx context.Context, filter model.TagFilter, limit, offset uint) ([]model.Tag, error)
 }
